--- client/src/pages/Page2.tsx
@@ -1,9 +1,0 @@
-const Page2: React.FC<{}> = () => {
-  return (
-    <div>
-      This second page is just for testing react router, you can delete when you have an actual 2nd page
-    </div>
-  );
-}
-
-export default Page2;
