+++ client/src/starpage/index.tsx
@@ -1,0 +1,9 @@
+const StarPage: React.FC<{}> = () => {
+  return (
+    <div>
+      StarPage goes here
+    </div>
+  );
+}
+
+export default StarPage;
